extract data fetching in helper function
